Zero PID error integral if there is any roll input
Unless the error integral is zeroed, the pilot will wind up fighting the
roll stabilizer while banking, and then due to the accumulated error,
the plane will start contorting wildly (and possibly catastrophically)
as the PID tries to correct the error.

The input threshold of 0.01 seems to work well even with a bit of slop
in a joystick's deadzone, but might need tuning.